+++ public.ssl/modules/actes/actes_transac_show.php
@@ -575,17 +575,22 @@ if ($me->isSuper()) {
 
 
 	if (in_array($transStatus,array(3,-1))  && $trans->get("type") == 1) {
+
 		$actionHtml .= "<form action=\"" . WEBSITE_SSL . "/modules/actes/actes_transac_rolback_attente.php\" onsubmit=\"return confirm('Êtes-vous certain de vouloir faire cela ? ')\" method=\"post\">\n";
-		$actionHtml .= "<div class=\"form-group\">\n<label class=\"col-md-4 control-label\">Revenir à l'état En attente de transmission </label>\n";
+		$actionHtml .= "<div class=\"form-group\">\n<label class=\"col-md-4 control-label\">Passer à En attente de transmission </label>\n";
 		$actionHtml .= "<input type=\"hidden\" name=\"id\" value=\"" . $trans->getId() . "\" />\n";
-		$actionHtml .= "<input type=\"submit\" value=\"Revenir en arrière\" class=\"btn btn-warning\" />\n";
+		$actionHtml .= "<input type=\"submit\" value=\"Passer en attente de transmission\" class=\"btn btn-warning\" />\n";
+		$actionHtml .= "</div></form>\n";
+
+		$actionHtml .= "<form action=\"" . WEBSITE_SSL . "/modules/actes/actes_transac_rolback_attente.php\" onsubmit=\"return confirm('Êtes-vous certain de vouloir faire cela ? ')\" method=\"post\">\n";
+		$actionHtml .= "<div class=\"form-group\">\n<label class=\"col-md-4 control-label\">Passer à Poster </label>\n";
+		$actionHtml .= "<input type=\"hidden\" name=\"id\" value=\"" . $trans->getId() . "\" />\n";
+		$actionHtml .= "<input type=\"hidden\" name=\"status_id\" value=\"1\" />\n";
+		$actionHtml .= "<input type=\"submit\" value=\"Passer à Poster\" class=\"btn btn-warning\" />\n";
 		$actionHtml .= "</div></form>\n";
 	}
 
-
 }
-
-
 
 
 if (isset($actionHtml)) {